requirements.txt
appdirs==1.4.4
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
backports.entry-points-selectable==1.1.1
black==21.7b0
certifi==2021.10.8
cfgv==3.3.1
chardet==3.0.4
click==8.0.3
colorama==0.4.4
distlib==0.3.3
dj-database-url==0.5.0
Django==3.0.3
django-cors-headers==3.7.0
django-filter==2.3.0
django-heroku==0.3.1
djangorestframework==3.11.1
filelock==3.4.0
gunicorn==20.0.4
identify==2.4.0
idna==2.10
iniconfig==1.1.1
isort==5.5.2
more-itertools==8.11.0
mypy-extensions==0.4.3
nodeenv==1.6.0
numpy==1.19.1
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.0
pluggy==0.13.1
pre-commit==2.15.0
psycopg2==2.9.2
py==1.11.0
pyparsing==3.0.6
pytest==6.0.1
pytest-django==3.9.0
python-decouple==3.4
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.24.0
six==1.15.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.2
urllib3==1.25.11
virtualenv==20.10.0
whitenoise==5.2.0
